+++ src/main/java/com/github/nthykier/debpkg/deb822/inspections/AbstractDctrlInspection.java
@@ -4,16 +4,20 @@ import com.github.nthykier.debpkg.deb822.dialects.Deb822DialectDebianControlLang
 import com.github.nthykier.debpkg.deb822.psi.Deb822File;
 import com.intellij.codeInspection.LocalInspectionTool;
 import com.intellij.codeInspection.ProblemsHolder;
+import com.intellij.lang.Language;
 import com.intellij.psi.PsiElementVisitor;
+import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 
+@RequiredArgsConstructor
 abstract class AbstractDctrlInspection extends LocalInspectionTool {
 
+    private final Language language;
 
     @NotNull
     public PsiElementVisitor buildVisitor(@NotNull final ProblemsHolder holder, final boolean isOnTheFly) {
         /* This only makes sense for debian/control files, but we are invoked for all Deb822 files */
-        if (holder.getFile().getLanguage() != Deb822DialectDebianControlLanguage.INSTANCE) {
+        if (!language.is(holder.getFile().getLanguage())) {
             return PsiElementVisitor.EMPTY_VISITOR;
         }
         if (!(holder.getFile() instanceof Deb822File)) {