+++ src/main/java/com/github/nthykier/debpkg/deb822/inspections/DCtrlMissingRulesRequiresRootInspection.java
@@ -7,6 +7,7 @@ import com.github.nthykier.debpkg.deb822.psi.Deb822Paragraph;
 import com.github.nthykier.debpkg.deb822.psi.Deb822Visitor;
 import com.intellij.codeInspection.LocalQuickFix;
 import com.intellij.codeInspection.ProblemsHolder;
+import com.intellij.lang.Language;
 import com.intellij.openapi.progress.ProgressIndicatorProvider;
 import com.intellij.psi.PsiElementVisitor;
 import org.jetbrains.annotations.NotNull;
@@ -14,6 +15,10 @@ import org.jetbrains.annotations.NotNull;
 import static com.github.nthykier.debpkg.util.AnnotatorUtil.fieldInsertionQuickFix;
 
 public class DCtrlMissingRulesRequiresRootInspection extends AbstractDctrlInspection {
+
+    public DCtrlMissingRulesRequiresRootInspection() {
+        super(Deb822DialectDebianControlLanguage.INSTANCE);
+    }
 
     @NotNull
     public PsiElementVisitor inspectionVisitor(@NotNull final ProblemsHolder holder, final boolean isOnTheFly) {