+++ src/main/java/com/github/nthykier/debpkg/deb822/inspections/DCtrlUsesDeprecatedImportantFieldWithoutProtectedInspection.java
@@ -1,11 +1,13 @@
 package com.github.nthykier.debpkg.deb822.inspections;
 
 import com.github.nthykier.debpkg.Deb822Bundle;
+import com.github.nthykier.debpkg.deb822.dialects.Deb822DialectDebianControlLanguage;
 import com.github.nthykier.debpkg.deb822.psi.Deb822FieldValuePair;
 import com.github.nthykier.debpkg.deb822.psi.Deb822Paragraph;
 import com.github.nthykier.debpkg.deb822.psi.Deb822Visitor;
 import com.intellij.codeInspection.LocalQuickFix;
 import com.intellij.codeInspection.ProblemsHolder;
+import com.intellij.lang.Language;
 import com.intellij.openapi.progress.ProgressIndicatorProvider;
 import com.intellij.psi.PsiElementVisitor;
 import org.jetbrains.annotations.NotNull;
@@ -13,6 +15,10 @@ import org.jetbrains.annotations.NotNull;
 import static com.github.nthykier.debpkg.util.AnnotatorUtil.fieldInsertionQuickFix;
 
 public class DCtrlUsesDeprecatedImportantFieldWithoutProtectedInspection extends AbstractDctrlInspection {
+
+    public DCtrlUsesDeprecatedImportantFieldWithoutProtectedInspection() {
+        super(Deb822DialectDebianControlLanguage.INSTANCE);
+    }
 
     @NotNull
     public PsiElementVisitor inspectionVisitor(@NotNull final ProblemsHolder holder, final boolean isOnTheFly) {