+++ CHANGELOG.md
@@ -11,9 +11,8 @@
 ## [Unreleased]
 ### Added
 - Recognise `debian/tests/control` automatically as a deb822 file
-- Recognise the standard fields and most common values for `debian/control/tests`. However, no `debian/control/tests`
-  specific validation has been added (e.g., the plugin not catch that `Tests` and `Test-Command` are mutually
-  exclusive).
+- Recognise the standard fields and most common values for `debian/control/tests`.
+- In `debian/tests/control`, show an error if `Tests` and `Test-Command` fields are used in the same paragraph.
 
 ### Fixed
 - Corrected display name of the inspection for non-standard Build-Profiles.